sendlist/page.tsxのコンポーネント分割
